```  rename `thermal_printer_plus_plugin.h` to `thermal_printer_plugin.h` ```
