+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: thermal_printer_plus
 description: A flutter plugin that prints esc commands to printers in different platforms such as android, ios, windows and different interfaces Bluetooth and BLE, TCP and USB
-version: 1.0.9
+version: 1.0.10
 homepage: https://github.com/nasibudesign/thermal_printer/
 
 # This package supports all platforms listed below.
